+++ src/files.d.ts
@@ -1,0 +1,4 @@
+interface FileInfo {
+  readonly name: string
+  creationTime: Date
+}