+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Only deploy after passing tests by using stages in Gitlab pipeline
 - Remove page headings on shots and settings pages
 - Update dependencies
+- Replace Vuex with Pinia state management
 
 ### Deprecated